# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Seha Sick Leave Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Leave {
  id: number;
  leaveCode: string;
  idNumber: string;
  patientNameAr: string;
  patientNameEn: string;
  /** @nullable */
  nationalityAr?: string | null;
  /** @nullable */
  nationalityEn?: string | null;
  /** @nullable */
  employerAr?: string | null;
  /** @nullable */
  employerEn?: string | null;
  doctorNameAr: string;
  doctorNameEn: string;
  /** @nullable */
  positionAr?: string | null;
  /** @nullable */
  positionEn?: string | null;
  admissionDateGregorian: string;
  /** @nullable */
  admissionDateHijri?: string | null;
  dischargeDateGregorian: string;
  /** @nullable */
  dischargeDateHijri?: string | null;
  issueDate: string;
  hospitalNameAr: string;
  hospitalNameEn: string;
  /** @nullable */
  reportTime?: string | null;
  /** @nullable */
  durationDays?: number | null;
  createdAt: string;
}

export interface LeaveInput {
  leaveCode: string;
  idNumber: string;
  patientNameAr: string;
  patientNameEn: string;
  nationalityAr?: string;
  nationalityEn?: string;
  employerAr?: string;
  employerEn?: string;
  doctorNameAr: string;
  doctorNameEn: string;
  positionAr?: string;
  positionEn?: string;
  admissionDateGregorian: string;
  admissionDateHijri?: string;
  dischargeDateGregorian: string;
  dischargeDateHijri?: string;
  issueDate: string;
  hospitalNameAr: string;
  hospitalNameEn: string;
  reportTime?: string;
  durationDays?: number;
}

export interface LeavesPage {
  data: Leave[];
  total: number;
  page: number;
  limit: number;
}

export interface LeaveStats {
  totalLeaves: number;
  todayLeaves: number;
  thisMonthLeaves: number;
  avgDuration: number;
  recentLeaves: Leave[];
}

export type ListLeavesParams = {
search?: string;
page?: number;
limit?: number;
};

export type InquireLeaveParams = {
idNumber: string;
leaveCode: string;
};