# lib/api-zod/src/generated/types/leave.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Seha Sick Leave Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Leave {
  id: number;
  leaveCode: string;
  idNumber: string;
  patientNameAr: string;
  patientNameEn: string;
  /** @nullable */
  nationalityAr?: string | null;
  /** @nullable */
  nationalityEn?: string | null;
  /** @nullable */
  employerAr?: string | null;
  /** @nullable */
  employerEn?: string | null;
  doctorNameAr: string;
  doctorNameEn: string;
  /** @nullable */
  positionAr?: string | null;
  /** @nullable */
  positionEn?: string | null;
  admissionDateGregorian: string;
  /** @nullable */
  admissionDateHijri?: string | null;
  dischargeDateGregorian: string;
  /** @nullable */
  dischargeDateHijri?: string | null;
  issueDate: string;
  hospitalNameAr: string;
  hospitalNameEn: string;
  /** @nullable */
  reportTime?: string | null;
  /** @nullable */
  durationDays?: number | null;
  createdAt: string;
}