# lib/api-zod/src/generated/types/leaveInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Seha Sick Leave Management API
 * OpenAPI spec version: 0.1.0
 */

export interface LeaveInput {
  leaveCode: string;
  idNumber: string;
  patientNameAr: string;
  patientNameEn: string;
  nationalityAr?: string;
  nationalityEn?: string;
  employerAr?: string;
  employerEn?: string;
  doctorNameAr: string;
  doctorNameEn: string;
  positionAr?: string;
  positionEn?: string;
  admissionDateGregorian: string;
  admissionDateHijri?: string;
  dischargeDateGregorian: string;
  dischargeDateHijri?: string;
  issueDate: string;
  hospitalNameAr: string;
  hospitalNameEn: string;
  reportTime?: string;
  durationDays?: number;
}