# lib/api-zod/src/generated/types/leaveStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Seha Sick Leave Management API
 * OpenAPI spec version: 0.1.0
 */
import type { Leave } from './leave';

export interface LeaveStats {
  totalLeaves: number;
  todayLeaves: number;
  thisMonthLeaves: number;
  avgDuration: number;
  recentLeaves: Leave[];
}