# lib/api-zod/src/generated/types/leavesPage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Seha Sick Leave Management API
 * OpenAPI spec version: 0.1.0
 */
import type { Leave } from './leave';

export interface LeavesPage {
  data: Leave[];
  total: number;
  page: number;
  limit: number;
}